# [rope] Faster tree builder