--- a/rust/rope/src/rope.rs
+++ b/rust/rope/src/rope.rs
@@ -578,7 +578,8 @@ impl TreeBuilder<RopeInfo> {
     /// Push a string on the accumulating tree in the naive way.
     ///
     /// Splits the provided string in chunks that fit in a leaf
-    /// and pushes the leaves one by one onto the tree by calling.
+    /// and pushes the leaves one by one onto the tree by calling
+    /// `push_leaf` on the builder.
     pub fn push_str(&mut self, mut s: &str) {
         if s.len() <= MAX_LEAF {
             if !s.is_empty() {
@@ -592,28 +593,6 @@ impl TreeBuilder<RopeInfo> {
             s = &s[splitpoint..];
         }
     }
-
-    /// Push a string on the accumulating tree in an optimized fashion.
-    ///
-    /// Splits the string into leaves first and
-    /// then pushes all the leaves onto the accumulating tree in one go.
-    ///
-    /// Note: this is only used in tests.
-    #[doc(hidden)]
-    pub fn push_str_stacked(&mut self, s: &str) {
-        let leaves = split_as_leaves(s);
-        self.push_leaves(leaves);
-    }
-}
-
-fn split_as_leaves(mut s: &str) -> Vec<String> {
-    let mut nodes = Vec::new();
-    while !s.is_empty() {
-        let splitpoint = if s.len() > MAX_LEAF { find_leaf_split_for_bulk(s) } else { s.len() };
-        nodes.push(s[..splitpoint].to_owned());
-        s = &s[splitpoint..];
-    }
-    nodes
 }
 
 impl<T: AsRef<str>> From<T> for Rope {