This patch is a rewrite of the TreeBuilder to be more aggressive in maintaining a stack of partially-built trees rather than basically delegating everything to a pairwise concat. In some cases (benchmarks to follow in a comment) the performance is significantly improved. I should note there are minor performance regressions in other cases; it would be good to investigate them in a followup, but I'm not especially concerned. More opportunities for optimization exist here, and my current goal is to get things "good enough."

Also, I found a case where the old code could generate trees violating the balancing invariants. This probably only affected performance, but was still not good. The PR contains debug asserts and a test.

Documentation is also improved.

- [ ] I have responded to reviews and made changes where appropriate.
- [x] I have tested the code with `cargo test --all` / `./rust/run_all_checks`.
- [x] I have updated comments / documentation related to the changes I made.
- [x] I have rebased my PR branch onto xi-editor/master.