@raphlinus 2020-09-08
Benchmark before:

```
running 3 tests
test benchmark_few_big_lines ... bench:       1,179 ns/iter (+/- 26)
test benchmark_short_lines   ... bench:  19,826,180 ns/iter (+/- 145,968)
test benchmark_triangle      ... bench:  41,039,930 ns/iter (+/- 306,473)

test result: ok. 0 passed; 0 failed; 0 ignored; 3 measured; 0 filtered out


running 10 tests
test hash_diff      ... bench:         925 ns/iter (+/- 24)
test hash_diff_big  ... bench:     142,166 ns/iter (+/- 19,248)
test hash_diff_med  ... bench:      23,873 ns/iter (+/- 430)
test ne_idx_avx     ... bench:       2,146 ns/iter (+/- 57)
test ne_idx_detect  ... bench:       2,153 ns/iter (+/- 312)
test ne_idx_rev_sse ... bench:       6,420 ns/iter (+/- 132)
test ne_idx_rev_sw  ... bench:      22,528 ns/iter (+/- 240)
test ne_idx_sse     ... bench:       4,239 ns/iter (+/- 90)
test ne_idx_sw      ... bench:           0 ns/iter (+/- 0)
test scanner        ... bench:       6,221 ns/iter (+/- 204)

test result: ok. 0 passed; 0 failed; 0 ignored; 10 measured; 0 filtered out


running 8 tests
test benchmark_char_insertion_one_line_edit ... bench:       1,681 ns/iter (+/- 65)
test benchmark_file_load_few_big_lines      ... bench:  20,923,440 ns/iter (+/- 960,330)
test benchmark_file_load_short_lines        ... bench:   6,204,325 ns/iter (+/- 190,156)
test benchmark_insert_newline               ... bench:       1,691 ns/iter (+/- 80)
test benchmark_overwrite_into_line          ... bench:      23,661 ns/iter (+/- 2,386)
test benchmark_paste_into_line              ... bench:       2,013 ns/iter (+/- 174)
test benchmark_triangle_concat_inplace      ... bench:   7,931,055 ns/iter (+/- 237,708)
test real_world_editing_scenario            ... bench:  31,900,450 ns/iter (+/- 179,303)
```

Benchmark after:

```
running 3 tests
test benchmark_few_big_lines ... bench:       1,169 ns/iter (+/- 42)
test benchmark_short_lines   ... bench:  19,265,220 ns/iter (+/- 156,707)
test benchmark_triangle      ... bench:  37,450,910 ns/iter (+/- 309,351)

test result: ok. 0 passed; 0 failed; 0 ignored; 3 measured; 0 filtered out


running 10 tests
test hash_diff      ... bench:       1,046 ns/iter (+/- 38)
test hash_diff_big  ... bench:     137,101 ns/iter (+/- 7,241)
test hash_diff_med  ... bench:      23,180 ns/iter (+/- 333)
test ne_idx_avx     ... bench:       2,143 ns/iter (+/- 28)
test ne_idx_detect  ... bench:       2,150 ns/iter (+/- 37)
test ne_idx_rev_sse ... bench:       6,401 ns/iter (+/- 80)
test ne_idx_rev_sw  ... bench:      22,535 ns/iter (+/- 248)
test ne_idx_sse     ... bench:       4,236 ns/iter (+/- 116)
test ne_idx_sw      ... bench:           0 ns/iter (+/- 0)
test scanner        ... bench:       6,020 ns/iter (+/- 229)

test result: ok. 0 passed; 0 failed; 0 ignored; 10 measured; 0 filtered out


running 8 tests
test benchmark_char_insertion_one_line_edit ... bench:       1,902 ns/iter (+/- 76)
test benchmark_file_load_few_big_lines      ... bench:  12,167,450 ns/iter (+/- 1,337,500)
test benchmark_file_load_short_lines        ... bench:   2,990,500 ns/iter (+/- 277,459)
test benchmark_insert_newline               ... bench:       2,032 ns/iter (+/- 63)
test benchmark_overwrite_into_line          ... bench:       6,496 ns/iter (+/- 266)
test benchmark_paste_into_line              ... bench:       2,188 ns/iter (+/- 117)
test benchmark_triangle_concat_inplace      ... bench:   4,545,190 ns/iter (+/- 194,200)
test real_world_editing_scenario            ... bench:  25,028,810 ns/iter (+/- 207,209)
```

The "overwrite_into_line" benchmark is nearly 4x faster. The file loading benchmarks are almost 2x faster, as is triangle_concat_inplace. The "real world editing" scenario is maybe 15% faster. As mentioned, there are a few small regressions. Tests that involve reading the ropes are unaffected. If I had constructed a benchmark that was sensitive to unbalanced ropes, and used `push_leaves` to make an unbalanced rope, I could have also shown a nice performance increase with that.